requirements.txt
flask
openai
python-dotenv
requests 
mcp-python